refactor: room 모듈 삭제
