+++ backend/src/lecture/dto/enter-lecture.dto.ts
@@ -1,0 +1,3 @@
+export class EnterLectureDto {
+  email: string;
+}
