--- backend/src/room/dto/create-room.dto.ts
@@ -1,5 +1,0 @@
-export class CreateRoomDto {
-  email: string;
-  title: string;
-  description: string;
-}
