--- backend/src/room/dto/enter-room.dto.ts
@@ -1,3 +1,0 @@
-export class EnterRoomDto {
-  email: string;
-}
