--- backend/src/room/room-code.schema.ts
@@ -1,14 +1,0 @@
-import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import mongoose from 'mongoose';
-import { Lecture } from '../lecture/lecture.schema';
-
-@Schema()
-export class EnterCode {
-  @Prop({ required: true })
-  code: string;
-
-  @Prop({ type: mongoose.Schema.Types.ObjectId, ref: 'Lecture' })
-  lecture_id: Lecture;
-}
-
-export const EnterCodeSchema = SchemaFactory.createForClass(EnterCode);
