--- backend/src/room/room.schema.ts
@@ -1,18 +1,0 @@
-import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { HydratedDocument } from 'mongoose';
-
-export type RoomDocument = HydratedDocument<Room>;
-
-@Schema()
-export class Room {
-  @Prop({ required: true })
-  title: string;
-
-  @Prop({ required: true })
-  description: string;
-
-  @Prop()
-  presenter_id: string;
-}
-
-export const RoomSchema = SchemaFactory.createForClass(Room);
